Move DataSet to it hown file Add Scatter on DataSet class Move  wheight updating to MLP (not so good) Add partial static Loss class that implement MSE Add "loss.Grad = 1" from program.cs to Value.Backpropagate. Add implicit cast from double to Value and an explicit from Value to double
